Move serialize back to meta namespace
