+++ py-polars/docs/source/reference/expressions/miscellaneous.rst
@@ -8,5 +8,4 @@ Miscellaneous
 
    Expr.deserialize
    Expr.from_json
-   Expr.serialize
    Expr.set_sorted
